requirements_4mac.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
audioread==3.1.0
av==12.3.0
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cn2an==0.5.22
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.2
ctranslate2==4.6.3
curl_cffi==0.13.0
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.108.0
faster-whisper==1.0.3
fastjsonschema==2.21.2
feedparser==6.0.12
ffmpeg==1.4
filelock==3.20.0
flatbuffers==25.12.19
fonttools==4.60.1
fqdn==1.5.1
frozendict==2.4.7
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.4.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-core==2.5.0
google-cloud-speech==2.26.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-generativeai==0.3.2
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.62.3
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.2.4
humanfriendly==10.0
idna==3.11
importlib-metadata==6.11.0
ipykernel==7.1.0
ipython==8.37.0
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiwer==3.0.3
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
lark==1.3.1
lazy_loader==0.4
Levenshtein==0.23.0
librosa==0.10.1
llvmlite==0.45.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.1.4
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multitasking==0.0.12
narwhals==2.13.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.5.0
notebook_shim==0.2.4
numba==0.62.1
numpy==1.26.4
onnxruntime==1.23.2
openai-whisper==20231117
opencc-python-reimplemented==0.1.7
opencv-python==4.8.1.78
overrides==7.7.0
packaging==23.2
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
peewee==3.18.3
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.5.0
pooch==1.8.2
proces==0.1.7
prometheus_client==0.23.1
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.1.3
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-json-logger==4.0.0
python-Levenshtein==0.23.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.14.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.9.4
rpds-py==0.29.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.12.2
Send2Trash==1.8.3
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.3
soundfile==0.12.1
soupsieve==2.8
soxr==1.0.0
SQLAlchemy==2.0.23
stack-data==0.6.3
starlette==0.32.0.post1
streamlit==1.29.0
svgwrite==1.4.3
sympy==1.14.0
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.22.2
toml==0.10.2
tomli==2.3.0
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
Tree==0.2.4
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.25.0
uvloop==0.22.1
validators==0.35.0
watchfiles==1.1.1
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
webrtcvad==2.0.10
websocket-client==1.9.0
websockets==16.0
widgetsnbextension==4.0.15
yfinance==1.0
zipp==3.23.0
